Adapt examples to be recognised as such by Cargo
Currently examples are binary modules inside the workspace. As a result,
these are not treated as examples by cargo, e.g.:

    > cargo run --example
    error: "--example" takes one argument.
    No examples available.

This patch moves files around a bit so that examples are actually
recognised as such by cargo. This has a few effects:

* `cargo test` will ensure that they compile.
* They will share the lock file with the main project. This ensure that
  their dependencies are aligned and avoids potentially downloading and
  compiling multiple different versions of the same dependency.
* `cargo run --example` provides useful output:

    > cargo run --example
    error: "--example" takes one argument.
    Available examples:
        editor-libcosmic
        editor-orbclient
        editor-test
        rich-text
        terminal

Depending on the shell's setup, shell completion may also be available
when using `cargo run --example`.